+++ modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/exec/rel/ExecutionTest.java
@@ -19,12 +19,9 @@ package org.apache.ignite.internal.sql.engine.exec.rel;
 
 import static java.lang.Math.max;
 import static java.lang.Math.min;
-import static org.apache.calcite.rel.core.JoinRelType.ANTI;
 import static org.apache.calcite.rel.core.JoinRelType.FULL;
 import static org.apache.calcite.rel.core.JoinRelType.INNER;
 import static org.apache.calcite.rel.core.JoinRelType.LEFT;
-import static org.apache.calcite.rel.core.JoinRelType.RIGHT;
-import static org.apache.calcite.rel.core.JoinRelType.SEMI;
 import static org.apache.ignite.internal.sql.engine.util.Commons.getFieldFromBiRows;
 import static org.apache.ignite.internal.testframework.IgniteTestUtils.assertThrowsWithCause;
 import static org.apache.ignite.internal.util.ArrayUtils.asList;
@@ -38,10 +35,8 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Stream;
 import org.apache.calcite.rel.core.CorrelationId;
-import org.apache.calcite.rel.core.JoinInfo;
 import org.apache.calcite.rel.core.JoinRelType;
 import org.apache.calcite.rel.type.RelDataType;
-import org.apache.calcite.util.ImmutableIntList;
 import org.apache.ignite.internal.sql.engine.exec.ExecutionContext;
 import org.apache.ignite.internal.sql.engine.exec.RowHandler;
 import org.apache.ignite.internal.sql.engine.exec.row.RowSchema;
@@ -57,16 +52,14 @@ import org.junit.jupiter.api.condition.OS;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 /**
  * ExecutionTest.
  * TODO Documentation https://issues.apache.org/jira/browse/IGNITE-15859
  */
 public class ExecutionTest extends AbstractExecutionTest<Object[]> {
-    @ParameterizedTest(name = "join algo : {0}")
-    @ValueSource(strings = {"NLJoin", "HashJoin"})
-    public void testSimpleExecution(String joinImplementation) {
+    @Test
+    public void testSimpleExecution() {
         // SELECT P.ID, P.NAME, PR.NAME AS PROJECT
         // FROM PERSON P
         // INNER JOIN PROJECT PR
@@ -100,16 +93,8 @@ public class ExecutionTest extends AbstractExecutionTest<Object[]> {
 
         RowHandler<Object[]> hnd = ctx.rowHandler();
 
-        AbstractRightMaterializedJoinNode<Object[]> join;
-
-        if ("NLJoin".equals(joinImplementation)) {
-            join = NestedLoopJoinNode.create(ctx, outType, leftType, rightType, INNER,
-                    (r1, r2) -> getFieldFromBiRows(hnd, 0, r1, r2) == getFieldFromBiRows(hnd, 4, r1, r2));
-        } else {
-            join = HashJoinNode.create(ctx, outType, leftType, rightType, INNER,
-                JoinInfo.of(ImmutableIntList.of(0), ImmutableIntList.of(1)));
-        }
-
+        NestedLoopJoinNode<Object[]> join = NestedLoopJoinNode.create(ctx, outType, leftType, rightType, INNER,
+                (r1, r2) -> getFieldFromBiRows(hnd, 0, r1, r2) == getFieldFromBiRows(hnd, 4, r1, r2));
         join.register(asList(persons, projects));
 
         ProjectNode<Object[]> project = new ProjectNode<>(ctx, r -> new Object[]{r[0], r[1], r[5]});
@@ -177,9 +162,8 @@ public class ExecutionTest extends AbstractExecutionTest<Object[]> {
         assertEquals(12, res.size());
     }
 
-    @ParameterizedTest(name = "join algo : {0}")
-    @ValueSource(strings = {"NLJoin", "HashJoin"})
-    public void testLeftJoin(String joinAlgo) {
+    @Test
+    public void testLeftJoin() {
         //    select e.id, e.name, d.name as dep_name
         //      from emp e
         // left join dep d
@@ -209,16 +193,8 @@ public class ExecutionTest extends AbstractExecutionTest<Object[]> {
 
         RowHandler<Object[]> hnd = ctx.rowHandler();
 
-        AbstractRightMaterializedJoinNode<Object[]> join;
-
-        if ("NLJoin".equals(joinAlgo)) {
-            join = NestedLoopJoinNode.create(ctx, outType, leftType, rightType, LEFT,
-                    (r1, r2) -> getFieldFromBiRows(hnd, 2, r1, r2) == getFieldFromBiRows(hnd, 3, r1, r2));
-        } else {
-            join = HashJoinNode.create(ctx, outType, leftType, rightType, LEFT,
-                    JoinInfo.of(ImmutableIntList.of(2), ImmutableIntList.of(0)));
-        }
-
+        NestedLoopJoinNode<Object[]> join = NestedLoopJoinNode.create(ctx, outType, leftType, rightType, LEFT,
+                (r1, r2) -> getFieldFromBiRows(hnd, 2, r1, r2) == getFieldFromBiRows(hnd, 3, r1, r2));
         join.register(asList(persons, deps));
 
         ProjectNode<Object[]> project = new ProjectNode<>(ctx, r -> new Object[]{r[0], r[1], r[4]});
@@ -237,73 +213,6 @@ public class ExecutionTest extends AbstractExecutionTest<Object[]> {
 
         assertEquals(4, rows.size());
 
-        assertArrayEquals(new Object[]{0, "Igor", "Core"}, rows.get(0));
-        assertArrayEquals(new Object[]{1, "Roman", "SQL"}, rows.get(1));
-        assertArrayEquals(new Object[]{2, "Ivan", null}, rows.get(2));
-        assertArrayEquals(new Object[]{3, "Alexey", "Core"}, rows.get(3));
-    }
-
-    @ParameterizedTest(name = "join algo : {0}")
-    @ValueSource(strings = {"NLJoin", "HashJoin"})
-    public void testRightJoin(String joinAlgo) {
-        //     select e.id, e.name, d.name as dep_name
-        //       from dep d
-        // right join emp e
-        //         on e.depno = d.depno
-
-        ExecutionContext<Object[]> ctx = executionContext(true);
-
-        ScanNode<Object[]> persons = new ScanNode<>(ctx, Arrays.asList(
-                new Object[]{0, "Igor", 1},
-                new Object[]{1, "Roman", 2},
-                new Object[]{2, "Ivan", null},
-                new Object[]{3, "Alexey", 1}
-        ));
-
-        ScanNode<Object[]> deps = new ScanNode<>(ctx, Arrays.asList(
-                new Object[]{1, "Core"},
-                new Object[]{2, "SQL"},
-                new Object[]{3, "QA"}
-        ));
-
-        IgniteTypeFactory tf = ctx.getTypeFactory();
-
-        RelDataType outType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-        RelDataType leftType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf, NativeTypes.INT32, NativeTypes.STRING));
-        RelDataType rightType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-
-        RowHandler<Object[]> hnd = ctx.rowHandler();
-
-        AbstractRightMaterializedJoinNode<Object[]> join;
-
-        if ("NLJoin".equals(joinAlgo)) {
-            join = NestedLoopJoinNode.create(ctx, outType, leftType, rightType, RIGHT,
-                    (r1, r2) -> getFieldFromBiRows(hnd, 0, r1, r2) == getFieldFromBiRows(hnd, 4, r1, r2));
-        } else {
-            join = HashJoinNode.create(ctx, outType, leftType, rightType, RIGHT,
-                    JoinInfo.of(ImmutableIntList.of(0), ImmutableIntList.of(2)));
-        }
-
-        join.register(asList(deps, persons));
-
-        ProjectNode<Object[]> project = new ProjectNode<>(ctx, r -> new Object[]{r[2], r[3], r[1]});
-        project.register(join);
-
-        RootNode<Object[]> node = new RootNode<>(ctx);
-        node.register(project);
-
-        assert node.hasNext();
-
-        ArrayList<Object[]> rows = new ArrayList<>();
-
-        while (node.hasNext()) {
-            rows.add(node.next());
-        }
-
-        assertEquals(4, rows.size());
-
         Object[][] expected = {
                 {0, "Igor", "Core"},
                 {3, "Alexey", "Core"},
@@ -314,91 +223,7 @@ public class ExecutionTest extends AbstractExecutionTest<Object[]> {
         assert2DimArrayEquals(expected, rows);
     }
 
-    @Test
-    public void testHashJoinRewind() {
-        ExecutionContext<Object[]> ctx = executionContext(true);
-
-        ScanNode<Object[]> persons = new ScanNode<>(ctx, Arrays.asList(
-                new Object[]{0, "Igor", 1},
-                new Object[]{1, "Roman", 2},
-                new Object[]{2, "Ivan", 5},
-                new Object[]{3, "Alexey", 1}
-        ));
-
-        ScanNode<Object[]> deps = new ScanNode<>(ctx, Arrays.asList(
-                new Object[]{1, "Core"},
-                new Object[]{2, "SQL"},
-                new Object[]{3, "QA"}
-        ));
-
-        IgniteTypeFactory tf = ctx.getTypeFactory();
-
-        RelDataType outType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-        RelDataType leftType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf, NativeTypes.INT32, NativeTypes.STRING));
-        RelDataType rightType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-
-        AbstractRightMaterializedJoinNode<Object[]> join = HashJoinNode.create(ctx, outType, leftType, rightType, RIGHT,
-                    JoinInfo.of(ImmutableIntList.of(0), ImmutableIntList.of(2)));
-
-        join.register(asList(deps, persons));
-
-        ProjectNode<Object[]> project = new ProjectNode<>(ctx, r -> new Object[]{r[2], r[3], r[1]});
-        project.register(join);
-
-        RootRewindable<Object[]> node = new RootRewindable<>(ctx);
-        node.register(project);
-
-        assert node.hasNext();
-
-        ArrayList<Object[]> rows = new ArrayList<>();
-
-        while (node.hasNext()) {
-            rows.add(node.next());
-        }
-
-        assertEquals(4, rows.size());
-
-        Object[][] expected = {
-                {0, "Igor", "Core"},
-                {3, "Alexey", "Core"},
-                {1, "Roman", "SQL"},
-                {2, "Ivan", null}
-        };
-
-        assert2DimArrayEquals(expected, rows);
-
-        List<Object[]> depsRes = new ArrayList<>();
-        depsRes.add(new Object[]{5, "QA"});
-
-        deps = new ScanNode<>(ctx, depsRes);
-
-        join.register(asList(deps, persons));
-
-        node.rewind();
-
-        assert node.hasNext();
-
-        ArrayList<Object[]> rowsAfterRewind = new ArrayList<>();
-
-        while (node.hasNext()) {
-            rowsAfterRewind.add(node.next());
-        }
-
-        assertEquals(4, rowsAfterRewind.size());
-
-        Object[][] expectedAfterRewind = {
-                {2, "Ivan", "QA"},
-                {1, "Roman", null},
-                {0, "Igor", null},
-                {3, "Alexey", null},
-        };
-
-        assert2DimArrayEquals(expectedAfterRewind, rowsAfterRewind);
-    }
-
-    private static void assert2DimArrayEquals(Object[][] expected, ArrayList<Object[]> actual) {
+    static void assert2DimArrayEquals(Object[][] expected, ArrayList<Object[]> actual) {
         assertEquals(expected.length, actual.size(), "expected length: " + expected.length + ", actual length: " + actual.size());
 
         int length = expected.length;
@@ -469,135 +294,6 @@ public class ExecutionTest extends AbstractExecutionTest<Object[]> {
         assertArrayEquals(new Object[]{2, "Ivan", null}, rows.get(2));
         assertArrayEquals(new Object[]{3, "Alexey", "Core"}, rows.get(3));
         assertArrayEquals(new Object[]{null, null, "QA"}, rows.get(4));
-    }
-
-    @ParameterizedTest(name = "join algo : {0}")
-    @ValueSource(strings = {"NLJoin", "HashJoin"})
-    public void testSemiJoin(String joinAlgo) {
-        //    select d.name as dep_name
-        //      from dep d
-        // semi join emp e
-        //        on e.depno = d.depno
-
-        ExecutionContext<Object[]> ctx = executionContext(true);
-
-        ScanNode<Object[]> persons = new ScanNode<>(ctx, Arrays.asList(
-                new Object[]{0, "Igor", 1},
-                new Object[]{1, "Roman", 2},
-                new Object[]{2, "Ivan", null},
-                new Object[]{3, "Alexey", 1}
-        ));
-
-        ScanNode<Object[]> deps = new ScanNode<>(ctx, Arrays.asList(
-                new Object[]{1, "Core"},
-                new Object[]{2, "SQL"},
-                new Object[]{3, "QA"}
-        ));
-
-        IgniteTypeFactory tf = ctx.getTypeFactory();
-
-        RelDataType outType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-        RelDataType leftType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-        RelDataType rightType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf, NativeTypes.INT32, NativeTypes.STRING));
-
-        RowHandler<Object[]> hnd = ctx.rowHandler();
-
-        AbstractRightMaterializedJoinNode<Object[]> join;
-
-        if ("NLJoin".equals(joinAlgo)) {
-            join = NestedLoopJoinNode.create(ctx, outType, leftType, rightType, SEMI,
-                    (r1, r2) -> getFieldFromBiRows(hnd, 0, r1, r2) == getFieldFromBiRows(hnd, 4, r1, r2));
-        } else {
-            join = HashJoinNode.create(ctx, outType, leftType, rightType, SEMI,
-                    JoinInfo.of(ImmutableIntList.of(0), ImmutableIntList.of(2)));
-        }
-
-        join.register(asList(deps, persons));
-
-        ProjectNode<Object[]> project = new ProjectNode<>(ctx, r -> new Object[]{r[1]});
-        project.register(join);
-
-        RootNode<Object[]> node = new RootNode<>(ctx);
-        node.register(project);
-
-        assert node.hasNext();
-
-        ArrayList<Object[]> rows = new ArrayList<>();
-
-        while (node.hasNext()) {
-            rows.add(node.next());
-        }
-
-        assertEquals(2, rows.size());
-
-        assertArrayEquals(new Object[]{"Core"}, rows.get(0));
-        assertArrayEquals(new Object[]{"SQL"}, rows.get(1));
-    }
-
-    @ParameterizedTest(name = "join algo : {0}")
-    @ValueSource(strings = {"NLJoin", "HashJoin"})
-    public void testAntiJoin(String joinAlgo) {
-        //    select d.name as dep_name
-        //      from dep d
-        // anti join emp e
-        //        on e.depno = d.depno
-
-        ExecutionContext<Object[]> ctx = executionContext(true);
-
-        ScanNode<Object[]> persons = new ScanNode<>(ctx, Arrays.asList(
-                new Object[]{0, "Igor", 1},
-                new Object[]{1, "Roman", 2},
-                new Object[]{2, "Ivan", null},
-                new Object[]{3, "Alexey", 1}
-        ));
-
-        ScanNode<Object[]> deps = new ScanNode<>(ctx, Arrays.asList(
-                new Object[]{1, "Core"},
-                new Object[]{2, "SQL"},
-                new Object[]{3, "QA"}
-        ));
-
-        IgniteTypeFactory tf = ctx.getTypeFactory();
-
-        RelDataType outType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-        RelDataType leftType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-        RelDataType rightType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf, NativeTypes.INT32, NativeTypes.STRING));
-
-        RowHandler<Object[]> hnd = ctx.rowHandler();
-
-        AbstractRightMaterializedJoinNode<Object[]> join;
-
-        if ("NLJoin".equals(joinAlgo)) {
-            join = NestedLoopJoinNode.create(ctx, outType, leftType, rightType, ANTI,
-                    (r1, r2) -> getFieldFromBiRows(hnd, 0, r1, r2) == getFieldFromBiRows(hnd, 4, r1, r2));
-        } else {
-            join = HashJoinNode.create(ctx, outType, leftType, rightType, ANTI,
-                    JoinInfo.of(ImmutableIntList.of(0), ImmutableIntList.of(2)));
-        }
-
-        join.register(asList(deps, persons));
-
-        ProjectNode<Object[]> project = new ProjectNode<>(ctx, r -> new Object[]{r[1]});
-        project.register(join);
-
-        RootNode<Object[]> node = new RootNode<>(ctx);
-        node.register(project);
-
-        assert node.hasNext();
-
-        ArrayList<Object[]> rows = new ArrayList<>();
-
-        while (node.hasNext()) {
-            rows.add(node.next());
-        }
-
-        assertEquals(1, rows.size());
-
-        assertArrayEquals(new Object[]{"QA"}, rows.get(0));
     }
 
     /**
