+++ modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/exec/rel/NestedLoopJoinExecutionTest.java
@@ -17,362 +17,10 @@
 
 package org.apache.ignite.internal.sql.engine.exec.rel;
 
-import static org.apache.calcite.rel.core.JoinRelType.ANTI;
-import static org.apache.calcite.rel.core.JoinRelType.FULL;
-import static org.apache.calcite.rel.core.JoinRelType.INNER;
-import static org.apache.calcite.rel.core.JoinRelType.LEFT;
-import static org.apache.calcite.rel.core.JoinRelType.RIGHT;
-import static org.apache.calcite.rel.core.JoinRelType.SEMI;
-import static org.apache.ignite.internal.sql.engine.util.Commons.getFieldFromBiRows;
-import static org.apache.ignite.internal.testframework.IgniteTestUtils.await;
-import static org.apache.ignite.internal.util.ArrayUtils.asList;
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
-import org.apache.calcite.rel.core.JoinRelType;
-import org.apache.calcite.rel.type.RelDataType;
-import org.apache.ignite.internal.sql.engine.exec.ExecutionContext;
-import org.apache.ignite.internal.sql.engine.exec.RowHandler;
-import org.apache.ignite.internal.sql.engine.exec.TestDownstream;
-import org.apache.ignite.internal.sql.engine.framework.ArrayRowHandler;
-import org.apache.ignite.internal.sql.engine.type.IgniteTypeFactory;
-import org.apache.ignite.internal.sql.engine.util.TypeUtils;
-import org.apache.ignite.internal.type.NativeTypes;
-import org.junit.jupiter.api.Test;
-
-/**
- * NestedLoopJoinExecutionTest.
- * TODO Documentation https://issues.apache.org/jira/browse/IGNITE-15859
- */
-public class NestedLoopJoinExecutionTest extends AbstractExecutionTest<Object[]> {
-    @Test
-    public void joinEmptyTables() {
-        verifyJoin(EMPTY, EMPTY, INNER, EMPTY);
-        verifyJoin(EMPTY, EMPTY, LEFT, EMPTY);
-        verifyJoin(EMPTY, EMPTY, RIGHT, EMPTY);
-        verifyJoin(EMPTY, EMPTY, FULL, EMPTY);
-        verifyJoin(EMPTY, EMPTY, SEMI, EMPTY);
-        verifyJoin(EMPTY, EMPTY, ANTI, EMPTY);
-    }
-
-    @Test
-    public void joinEmptyLeftTable() {
-        Object[][] right = {
-                {1, "Core"},
-                {1, "OLD_Core"},
-                {2, "SQL"}
-        };
-
-        verifyJoin(EMPTY, right, INNER, EMPTY);
-        verifyJoin(EMPTY, right, LEFT, EMPTY);
-        verifyJoin(EMPTY, right, RIGHT, new Object[][]{
-                {null, null, "Core"},
-                {null, null, "OLD_Core"},
-                {null, null, "SQL"}
-        });
-        verifyJoin(EMPTY, right, FULL, new Object[][]{
-                {null, null, "Core"},
-                {null, null, "OLD_Core"},
-                {null, null, "SQL"}
-        });
-        verifyJoin(EMPTY, right, SEMI, EMPTY);
-        verifyJoin(EMPTY, right, ANTI, EMPTY);
-    }
-
-    @Test
-    public void joinEmptyRightTable() {
-        Object[][] left = {
-                {1, "Roman", null},
-                {2, "Igor", 1},
-                {3, "Alexey", 2}
-        };
-
-        verifyJoin(left, EMPTY, INNER, EMPTY);
-        verifyJoin(left, EMPTY, LEFT, new Object[][]{
-                {1, "Roman", null},
-                {2, "Igor", null},
-                {3, "Alexey", null}
-        });
-        verifyJoin(left, EMPTY, RIGHT, EMPTY);
-        verifyJoin(left, EMPTY, FULL, new Object[][]{
-                {1, "Roman", null},
-                {2, "Igor", null},
-                {3, "Alexey", null}
-        });
-        verifyJoin(left, EMPTY, SEMI, EMPTY);
-        verifyJoin(left, EMPTY, ANTI, new Object[][]{
-                {1, "Roman"},
-                {2, "Igor"},
-                {3, "Alexey"}
-        });
-    }
-
-    @Test
-    public void joinOneToMany() {
-        Object[][] left = {
-                {1, "Roman", null},
-                {2, "Igor", 1},
-                {3, "Alexey", 2}
-        };
-
-        Object[][] right = {
-                {1, "Core"},
-                {1, "OLD_Core"},
-                {2, "SQL"},
-                {3, "Arch"}
-        };
-
-        verifyJoin(left, right, INNER, new Object[][]{
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Alexey", "SQL"}
-        });
-        verifyJoin(left, right, LEFT, new Object[][]{
-                {1, "Roman", null},
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Alexey", "SQL"}
-        });
-        verifyJoin(left, right, RIGHT, new Object[][]{
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Alexey", "SQL"},
-                {null, null, "Arch"}
-        });
-        verifyJoin(left, right, FULL, new Object[][]{
-                {1, "Roman", null},
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Alexey", "SQL"},
-                {null, null, "Arch"}
-        });
-        verifyJoin(left, right, SEMI, new Object[][]{
-                {2, "Igor"},
-                {3, "Alexey"}
-        });
-        verifyJoin(left, right, ANTI, new Object[][]{
-                {1, "Roman"}
-        });
-    }
-
-    @Test
-    public void joinOneToMany2() {
-        Object[][] left = {
-                {1, "Roman", null},
-                {2, "Igor", 1},
-                {3, "Alexey", 2},
-                {4, "Ivan", 4},
-                {5, "Taras", 5},
-                {6, "Lisa", 6}
-        };
-
-        Object[][] right = {
-                {1, "Core"},
-                {1, "OLD_Core"},
-                {2, "SQL"},
-                {3, "QA"},
-                {5, "Arch"}
-        };
-
-        verifyJoin(left, right, INNER, new Object[][]{
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Alexey", "SQL"},
-                {5, "Taras", "Arch"}
-        });
-        verifyJoin(left, right, LEFT, new Object[][]{
-                {1, "Roman", null},
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Alexey", "SQL"},
-                {4, "Ivan", null},
-                {5, "Taras", "Arch"},
-                {6, "Lisa", null}
-        });
-        verifyJoin(left, right, RIGHT, new Object[][]{
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Alexey", "SQL"},
-                {5, "Taras", "Arch"},
-                {null, null, "QA"}
-        });
-        verifyJoin(left, right, FULL, new Object[][]{
-                {1, "Roman", null},
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Alexey", "SQL"},
-                {4, "Ivan", null},
-                {5, "Taras", "Arch"},
-                {6, "Lisa", null},
-                {null, null, "QA"}
-        });
-        verifyJoin(left, right, SEMI, new Object[][]{
-                {2, "Igor"},
-                {3, "Alexey"},
-                {5, "Taras"}
-        });
-        verifyJoin(left, right, ANTI, new Object[][]{
-                {1, "Roman"},
-                {4, "Ivan"},
-                {6, "Lisa"}
-        });
-    }
-
-    @Test
-    public void joinManyToMany() {
-        Object[][] left = {
-                {1, "Roman", null},
-                {2, "Igor", 1},
-                {3, "Taras", 1},
-                {4, "Alexey", 2},
-                {5, "Ivan", 4},
-                {6, "Andrey", 4}
-        };
-
-        Object[][] right = {
-                {1, "Core"},
-                {1, "OLD_Core"},
-                {2, "SQL"},
-                {3, "Arch"},
-                {4, "QA"},
-                {4, "OLD_QA"},
-        };
-
-        verifyJoin(left, right, INNER, new Object[][]{
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Taras", "Core"},
-                {3, "Taras", "OLD_Core"},
-                {4, "Alexey", "SQL"},
-                {5, "Ivan", "QA"},
-                {5, "Ivan", "OLD_QA"},
-                {6, "Andrey", "QA"},
-                {6, "Andrey", "OLD_QA"}
-        });
-        verifyJoin(left, right, LEFT, new Object[][]{
-                {1, "Roman", null},
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Taras", "Core"},
-                {3, "Taras", "OLD_Core"},
-                {4, "Alexey", "SQL"},
-                {5, "Ivan", "QA"},
-                {5, "Ivan", "OLD_QA"},
-                {6, "Andrey", "QA"},
-                {6, "Andrey", "OLD_QA"}
-        });
-        verifyJoin(left, right, RIGHT, new Object[][]{
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Taras", "Core"},
-                {3, "Taras", "OLD_Core"},
-                {4, "Alexey", "SQL"},
-                {5, "Ivan", "QA"},
-                {5, "Ivan", "OLD_QA"},
-                {6, "Andrey", "QA"},
-                {6, "Andrey", "OLD_QA"},
-                {null, null, "Arch"}
-        });
-        verifyJoin(left, right, FULL, new Object[][]{
-                {1, "Roman", null},
-                {2, "Igor", "Core"},
-                {2, "Igor", "OLD_Core"},
-                {3, "Taras", "Core"},
-                {3, "Taras", "OLD_Core"},
-                {4, "Alexey", "SQL"},
-                {5, "Ivan", "QA"},
-                {5, "Ivan", "OLD_QA"},
-                {6, "Andrey", "QA"},
-                {6, "Andrey", "OLD_QA"},
-                {null, null, "Arch"}
-        });
-        verifyJoin(left, right, SEMI, new Object[][]{
-                {2, "Igor"},
-                {3, "Taras"},
-                {4, "Alexey"},
-                {5, "Ivan"},
-                {6, "Andrey"},
-        });
-        verifyJoin(left, right, ANTI, new Object[][]{
-                {1, "Roman"}
-        });
-    }
-
-    /**
-     * Creates execution tree and executes it. Then compares the result of the execution with the given one.
-     *
-     * @param left     Data for left table.
-     * @param right    Data for right table.
-     * @param joinType Join type.
-     * @param expRes   Expected result.
-     */
-    private void verifyJoin(Object[][] left, Object[][] right, JoinRelType joinType, Object[][] expRes) {
-        ExecutionContext<Object[]> ctx = executionContext(true);
-
-        IgniteTypeFactory tf = ctx.getTypeFactory();
-
-        RelDataType leftType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-        ScanNode<Object[]> leftNode = new ScanNode<>(ctx, Arrays.asList(left));
-
-        RelDataType rightType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf, NativeTypes.INT32, NativeTypes.STRING));
-        ScanNode<Object[]> rightNode = new ScanNode<>(ctx, Arrays.asList(right));
-
-        RelDataType outType;
-        if (setOf(SEMI, ANTI).contains(joinType)) {
-            outType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                    NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32));
-        } else {
-            outType = TypeUtils.createRowType(tf, TypeUtils.native2relationalTypes(tf,
-                    NativeTypes.INT32, NativeTypes.STRING, NativeTypes.INT32, NativeTypes.INT32, NativeTypes.STRING));
-        }
-
-        RowHandler<Object[]> hnd = ctx.rowHandler();
-
-        NestedLoopJoinNode<Object[]> join = NestedLoopJoinNode.create(ctx, outType, leftType, rightType, joinType,
-                (r1, r2) -> getFieldFromBiRows(hnd, 2, r1, r2) == getFieldFromBiRows(hnd, 3, r1, r2));
-        join.register(asList(leftNode, rightNode));
-
-        ProjectNode<Object[]> project;
-        if (setOf(SEMI, ANTI).contains(joinType)) {
-            project = new ProjectNode<>(ctx, r -> new Object[]{r[0], r[1]});
-        } else {
-            project = new ProjectNode<>(ctx, r -> new Object[]{r[0], r[1], r[4]});
-        }
-        project.register(join);
-
-        // first, let's rewind just created tree -- it's how it actually being executed
-        project.rewind();
-
-        int times = 2;
-        do {
-            TestDownstream<Object[]> downstream = new TestDownstream<>();
-            project.onRegister(downstream);
-
-            ctx.execute(() -> project.request(1024), project::onError);
-
-            assertArrayEquals(expRes, await(downstream.result()).toArray(EMPTY));
-
-            // now let's rewind and restart test to check whether all state has been correctly reset
-            project.rewind();
-        } while (times-- > 0);
-    }
-
-    /**
-     * Creates {@link Set set} from provided items.
-     *
-     * @param items Items.
-     * @return New set.
-     */
-    @SafeVarargs
-    private static <T> Set<T> setOf(T... items) {
-        return new HashSet<>(Arrays.asList(items));
-    }
-
+/** Nested loop join execution tests. */
+public class NestedLoopJoinExecutionTest extends AbstractJoinExecutionTest {
     @Override
-    protected RowHandler<Object[]> rowHandler() {
-        return ArrayRowHandler.INSTANCE;
+    JoinAlgo joinAlgo() {
+        return JoinAlgo.NESTED_LOOP;
     }
 }
